+++ U232/HgComplexTestSuite/Src/TestDynamicMessageSuite.h
@@ -450,7 +450,7 @@ void TestDynamicMessageSuite::TestCopyCtor()
 
   // SUT
   SUT sut(rhs);
-  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), sut.size());
+  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), (unsigned int)sut.size());
 }
 
 //  ****************************************************************************
@@ -460,7 +460,7 @@ void TestDynamicMessageSuite::TestValueCtor()
   PopulateBaseValues (rhs);
   // SUT
   SUT sut(rhs.data(), rhs.size());
-  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), sut.size());
+  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), (unsigned int)sut.size());
 }
 
 //  ****************************************************************************
@@ -472,7 +472,7 @@ void TestDynamicMessageSuite::TestAssignmentOperator()
   // SUT
   SUT sut;
   sut = rhs;
-  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), sut.size());
+  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), (unsigned int)sut.size());
 }
 
 //  ****************************************************************************
@@ -529,7 +529,7 @@ void TestDynamicMessageSuite::TestAssign()
   sut.assign(expected.data(), expected.size());
 
   TS_ASSERT_EQUALS(expected.size(), sut.size());
-  TS_ASSERT_SAME_DATA(expected.data(), sut.data(), sut.size());
+  TS_ASSERT_SAME_DATA(expected.data(), sut.data(), (unsigned int)sut.size());
 }
 
 //  ****************************************************************************
@@ -583,7 +583,7 @@ void TestDynamicMessageSuite::Testdata()
   sut.seq_64[2] = seq_64[2];
 
   TS_ASSERT_EQUALS(k_sut_msg_size, sut.size());
-  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), sut.size());
+  TS_ASSERT_SAME_DATA(packed_msg, sut.data(), (unsigned int)sut.size());
 }
 
 //  ****************************************************************************
@@ -606,8 +606,8 @@ void TestDynamicMessageSuite::Testto_host()
   SUT result = to_host(sut);
   SUT no_op_result = to_host(no_op_sut);
 
-  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), sut.size());
-  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), no_op_sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), (unsigned int)sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), (unsigned int)no_op_sut.size());
 }
 
 //  ****************************************************************************
@@ -630,8 +630,8 @@ void TestDynamicMessageSuite::Testto_network()
   SUT_net result = to_network(sut);
   SUT_net no_op_result = to_network(no_op_sut);
 
-  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), sut.size());
-  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), no_op_sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), (unsigned int)sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), (unsigned int)no_op_sut.size());
 }
 
 //  ****************************************************************************
@@ -654,8 +654,8 @@ void TestDynamicMessageSuite::Testto_big_endian()
   SUT_big result = to_big_endian(sut);
   SUT_big no_op_result = to_big_endian(no_op_sut);
 
-  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), sut.size());
-  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), no_op_sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), (unsigned int)sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), (unsigned int)no_op_sut.size());
 }
 
 //  ****************************************************************************
@@ -678,8 +678,8 @@ void TestDynamicMessageSuite::Testto_little_endian()
   SUT_little result        = to_little_endian(sut);
   SUT_little no_op_result  = to_little_endian(no_op_sut);
 
-  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), sut.size());
-  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), no_op_sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, result.data(), (unsigned int)sut.size());
+  TS_ASSERT_SAME_DATA(other_packed_msg, no_op_result.data(), (unsigned int)no_op_sut.size());
 }
